DAOS-3800 vos: allow PMDK tx to cover multiple modifications
One distributed transaction may modify multiple objects/akeys/dkeys
that reside on the same daos target. Each of them will be handled
via one VOS API call. It is expected that all these modificatinos
are covered by single PMDK transaction. It is good for efficiency,
and also simplify the work if some of the modification failed and
need to cancel all modifications.

This patch allows the upper layer caller (object) to notify the VOS
when to start/stop the PMDK transaction without exporting related
PMDK APIs outside of VOS.

Signed-off-by: Fan Yong <fan.yong@intel.com>